# api/openapi/internal-v1.openapi.yaml
openapi: 3.1.0
info:
  title: AFSCP Internal API
  version: 1.0.0
  description: Product-agnostic internal storage control plane API.
  license:
    name: Proprietary internal
    identifier: LicenseRef-Proprietary
servers:
  - url: https://afscp.internal
security:
  - serviceAuth: []
components:
  securitySchemes:
    serviceAuth:
      type: http
      scheme: bearer
      description: Deployment service credential or mTLS-derived bearer facade.
  parameters:
...
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      schema:
        type: string
...
    CallerService:
      name: X-AFSCP-Caller-Service
      in: header
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorEnvelope:
      $ref: ../schemas/afscp-internal-v1.schema.json#/$defs/ErrorEnvelope
...
    OperationRecord:
      $ref: ../schemas/afscp-internal-v1.schema.json#/$defs/OperationRecord
...
  responses:
    Error:
      description: Standard error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
paths:
...
  /internal/v1/operations/{operationId}:
    get:
      operationId: getOperation
      summary: Get a redacted operation record.
      description: >
        Operation inspection does not require X-AFSCP-Namespace-Id because
        stored operation.namespace_id may be null. The handler resolves the
        record by operationId, then enforces namespace authorization when the
        stored namespace_id is present or operator/global authorization when it
        is absent.
      parameters:
        - $ref: "#/components/parameters/CorrelationId"
        - $ref: "#/components/parameters/CallerService"
        - name: operationId
          in: path
          required: true
          schema:
            $ref: ../schemas/afscp-internal-v1.schema.json#/$defs/OperationId
      responses:
        "400":
          $ref: "#/components/responses/Error"
        "200":
          description: Redacted operation record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationRecord"
        default:
          $ref: "#/components/responses/Error"
...